fix: correctness, numerics, and FFI-safety fixes from a first-principles audit
- pls: use the W(PᵀW)⁻¹ rotation in transform so multi-component scores are correct
  (raw NIPALS weights were only valid for the first component)
- pid-python: reject non-C-contiguous numpy arrays instead of silently transposing;
  map PidError to ValueError / RuntimeError / NotImplementedError
- pid-runlog: hash the full ordered event sequence for trace_hash/--compare so distinct
  traces with the same final state no longer collide; add config-hash and bridge-causality checks
- pipeline: validate bootstrap_pid3 alpha (was an out-of-range index panic); pls_cv parsimony
  + all-folds-failed error; screen_pid2 upfront validation
- geometry: error on non-finite Gromov distances instead of returning Ok(NaN)
- invariants: NaN-guard the degree ratios for tiny/non-positive joint MI
- discrete_pid: scale-relative constant-column detection; report the actual mismatching row count
- preprocess: width-independent HashProjector bucketing (32-bit reproducibility)
- exp0: scale-aware monotonicity/invariant tolerances (1e-9 flagged kNN estimator noise);
  treat NaN r-bar/v-bar as N/A; validate --block-size
- add regression tests for the PLS rotation, bootstrap alpha, trace-hash collision, and the
  Python contiguity/error-mapping fixes

All gates pass: fmt, clippy (default + parallel), rustdoc -D warnings, 122 workspace tests,
102 parallel-feature tests, and the maturin wheel + pytest suite.

+++ crates/pid-core/src/preprocess.rs
@@ -122,7 +122,10 @@ impl HashProjector {
         let mut sign = Vec::with_capacity(in_dim);
         for j in 0..in_dim {
             let h = splitmix64_hash(seed, j as u64);
-            index.push((h as usize) % out_dim);
+            // Reduce modulo `out_dim` in u64 BEFORE narrowing to usize. `h as usize`
+            // truncates to 32 bits on 32-bit targets, which would make the documented,
+            // seed-reproducible bucketing platform-dependent.
+            index.push((h % out_dim as u64) as usize);
             sign.push(if (h & 1) == 0 { 1.0 } else { -1.0 });
         }
